requirements.txt
psycopg2-binary==2.9.10
 asgiref==3.8.1
 django==5.1.4
 sqlparse==0.5.2
djangorestframework==3.15.2